requirements.txt
flask==3.0.0
python-dotenv==1.0.0
replicate==0.22.0
gunicorn==21.2.0
pyngrok==7.0.5
pandas==2.2.3
numpy==2.0.1
requests-oauthlib>=1.3.1
six>=1.16.0
psycopg2-binary>=2.9.10
spacy
Flask-Cors>=4.0.0
joblib>=1.4.2
scikit-learn>=1.5.1
pydantic>=2.0.0
flasgger
